application update

+++ src/main/java/io/codelex/loan/microlending/InterestFactorService.java
@@ -3,6 +3,6 @@ package io.codelex.loan.microlending;
 import java.math.BigDecimal;
 
 public interface InterestFactorService {
-    BigDecimal extendLoanInterestFactor(Long amount, Long term);
+    BigDecimal extendLoanInterestFactor(BigDecimal amount, Integer term);
 
 }
